+++ packages/charting/src/types/IDataPoint.ts
@@ -110,3 +110,37 @@ export interface IChartProps {
    */
   lineChartData?: ILineChartPoints[];
 }
+
+export interface IVSChartDataPoint {
+  /**
+   * data the datapoint in the chart
+   */
+  data: number;
+
+  /**
+   * Legend text for the datapoint in the chart
+   */
+  legend: string;
+
+  /**
+   * color for the legend in the chart
+   */
+  color?: string;
+}
+
+export interface IVerticalStackedChartProps {
+  /**
+   * data for the points in the chart
+   */
+  chartData: IVSChartDataPoint[];
+
+  /**
+   * Data for x axis label for multistacked Vertical bar chart
+   */
+  xAxisPoint: number | string;
+
+  /**
+   * chart title for the chart
+   */
+  chartTitle?: string;
+}