+++ packages/fluentui/react-northstar/test/specs/components/Toolbar/ToolbarMenuItem-test.ts
@@ -1,12 +1,17 @@
-import { isConformant } from 'test/specs/commonTests';
+import { implementsShorthandProp, isConformant } from 'test/specs/commonTests';
 
-import ToolbarMenuItem from 'src/components/Toolbar/ToolbarMenuItem';
 import Box from 'src/components/Box/Box';
+import ToolbarMenu from 'src/components/Toolbar/ToolbarMenu';
+import ToolbarMenuItem from 'src/components/Toolbar/ToolbarMenuItem';
 
 describe('ToolbarMenuItem', () => {
   isConformant(ToolbarMenuItem, {
     wrapperComponent: Box,
     autoControlledProps: ['menuOpen'],
     constructorName: 'ToolbarMenuItem',
+  });
+  implementsShorthandProp(ToolbarMenuItem)('menu', ToolbarMenu, {
+    implementsPopper: true,
+    requiredProps: { menuOpen: true },
   });
 });