+++ packages/styling/src/utilities/icons.ts
@@ -192,7 +192,7 @@ export function getIcon(name?: string): IIconRecord | undefined {
       // tslint:disable-next-line:deprecation
       if (!options.disableWarnings && options.warnOnMissingIcons) {
         warn(
-          `The icon "${name}" was used but not registered. See http://aka.ms/fabric-icon-usage for more information.`,
+          `The icon "${name}" was used but not registered. See https://github.com/microsoft/fluentui/wiki/Using-icons for more information.`,
         );
       }
     }